ENH: Remove circular module dependancy
The itk/__init.py__ depends on itkExtras.py which
in turn depends on itk/__init.py.

Remove those circular dependancies by directly
importing from primary modules.

Introduce new file to more cleanly break circular dependencies

The itkInitHelpers.py is dedicated to functions needed
during the initialization of the itk module.
