+++ Wrapping/Generators/Python/itkExtras.py
@@ -24,14 +24,7 @@ import numpy
 
 from itkTemplate import output
 
-# The following line defines an ascii string used for dynamically refreshing
-# the import and progress callbacks on the same terminal line.
-# See http://www.termsys.demon.co.uk/vtansi.htm
-# \033 is the C-style octal code for an escape character
-# [2000D moves the cursor back 2000 columns, this is a brute force way of
-# getting to the start of the line.
-# [K erases the end of the line
-clrLine = "\033[2000D\033[K"
+from sys import stderr as system_error_stream
 
 
 def set_nthreads(number_of_threads):
@@ -59,90 +52,6 @@ def get_nthreads():
 
     threader = itk.MultiThreaderBase.New()
     return threader.GetGlobalDefaultNumberOfThreads()
-
-
-def auto_not_in_place(v=True):
-    """Force it to not run in place
-    """
-    import itkConfig
-
-    itkConfig.NotInPlace = v
-
-
-def auto_progress(progress_type=1):
-    """Set up auto progress report
-
-    progress_type:
-        1 or True -> auto progress be used in a terminal
-        2 -> simple auto progress (without special characters)
-        0 or False -> disable auto progress
-    """
-    import itkConfig
-
-    if progress_type is True or progress_type == 1:
-        itkConfig.ImportCallback = terminal_import_callback
-        itkConfig.ProgressCallback = terminal_progress_callback
-
-    elif progress_type == 2:
-        itkConfig.ImportCallback = simple_import_callback
-        itkConfig.ProgressCallback = simple_progress_callback
-
-    elif progress_type is False or progress_type == 0:
-        itkConfig.ImportCallback = None
-        itkConfig.ProgressCallback = None
-
-    else:
-        raise ValueError("Invalid auto progress type: " + repr(progress_type))
-
-
-def terminal_progress_callback(name, p):
-    """Display the progress of an object and clean the display once complete
-
-    This function can be used with itkConfig.ProgressCallback
-    """
-    print(clrLine + "%s: %f" % (name, p), file=system_error_stream, end="")
-    if p == 1:
-        print(clrLine, file=system_error_stream, end="")
-
-
-def terminal_import_callback(name, p):
-    """Display the loading of a module and clean the display once complete
-
-    This function can be used with itkConfig.ImportCallback
-    """
-    print(clrLine + "Loading %s... " % name, file=system_error_stream, end="")
-    if p == 1:
-        print(clrLine, file=system_error_stream, end="")
-
-
-def simple_import_callback(name, p):
-    """Print a message when a module is loading
-
-    This function can be used with itkConfig.ImportCallback
-    """
-    if p == 0:
-        print("Loading %s... " % name, file=system_error_stream, end="")
-    elif p == 1:
-        print("done", file=system_error_stream)
-
-
-def simple_progress_callback(name, p):
-    """Print a message when an object is running
-
-    This function can be used with itkConfig.ProgressCallback
-    """
-    if p == 0:
-        print("Running %s... " % name, file=system_error_stream, end="")
-    elif p == 1:
-        print("done", file=system_error_stream)
-
-
-def force_load():
-    """force itk to load all the submodules"""
-    import itk
-
-    for k in dir(itk):
-        getattr(itk, k)
 
 
 def echo(obj, f=system_error_stream):
@@ -255,8 +164,7 @@ def _get_itk_pixelid(numpy_array_type):
 
 
 def _GetArrayFromImage(image_or_filter, function, keep_axes, update):
-    """Get an Array with the content of the image buffer
-    """
+    """Get an Array with the content of the image buffer"""
     # Finds the image type
     import itk
 
@@ -269,8 +177,7 @@ def _GetArrayFromImage(image_or_filter, function, keep_axes, update):
 
 
 def GetArrayFromImage(image_or_filter, keep_axes=False, update=True):
-    """Get an array with the content of the image buffer
-    """
+    """Get an array with the content of the image buffer"""
     return _GetArrayFromImage(image_or_filter, "GetArrayFromImage", keep_axes, update)
 
 
@@ -278,8 +185,7 @@ array_from_image = GetArrayFromImage
 
 
 def GetArrayViewFromImage(image_or_filter, keep_axes=False, update=True):
-    """Get an array view with the content of the image buffer
-    """
+    """Get an array view with the content of the image buffer"""
     return _GetArrayFromImage(
         image_or_filter, "GetArrayViewFromImage", keep_axes, update
     )
@@ -1038,7 +944,7 @@ class templated_class:
         # extract the types from the arguments to instantiate the class
         import itk
 
-        types = tuple(itk.class_(o) for o in args)
+        types = tuple(class_(o) for o in args)
         return self[types].New(*args, **kargs)
 
     def __getitem__(self, template_parameters):
